requirements.txt
psycopg2-binary
requests
schedule
pyqt5
